refactor store to use shared interface, custom S3Store structs/opts, and wire together...
