+++ Makefile
@@ -1,3 +1,13 @@
+PANTRI_BIN := $(shell bazel cquery :pantri_but_go --output=files 2>/dev/null)
+
+build: bazel_build
+
+bazel_build: gazelle
+	bazel build :pantri_but_go
+	@echo Copy, past, and execute this in your shell for convenience:
+	@echo
+	@echo PANTRI_BIN=$(PANTRI_BIN)
+
 ci: compile_and_run lint
 
 compile_and_run:
@@ -7,3 +17,9 @@ compile_and_run:
 lint:
 	docker build --tag pantributgolint -f _ci/lint/Dockerfile .
 	docker run pantributgolint
+
+gazelle:
+	bazel run :gazelle
+
+minio:
+	docker run -p 9000:9000 -p 9001:9001 quay.io/minio/minio server /data --console-address ":9001"
