+++ internal/metadata/BUILD.bazel
@@ -3,8 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "metadata",
     srcs = ["metadata.go"],
-    importpath = "github.com/discentem/pantri_but_go/metadata",
-    visibility = ["//visibility:public"],
+    importpath = "github.com/discentem/pantri_but_go/internal/metadata",
+    visibility = ["//:__subpackages__"],
     deps = ["@com_github_spf13_afero//:afero"],
 )
