+++ internal/pantri/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "pantri",
     srcs = ["pantri.go"],
-    importpath = "github.com/discentem/pantri_but_go/pantri",
-    visibility = ["//visibility:public"],
+    importpath = "github.com/discentem/pantri_but_go/internal/pantri",
+    visibility = ["//:__subpackages__"],
     deps = [
-        "//stores",
+        "//internal/stores",
         "@com_github_google_logger//:logger",
         "@com_github_mitchellh_go_homedir//:go-homedir",
         "@com_github_spf13_afero//:afero",
@@ -15,11 +15,10 @@ go_library(
 
 go_test(
     name = "pantri_test",
-    size = "small",
-    timeout = "short",
     srcs = ["pantri_test.go"],
     embed = [":pantri"],
     deps = [
+        "//internal/stores",
         "@com_github_spf13_afero//:afero",
         "@com_github_stretchr_testify//assert",
     ],
