+++ internal/pantri/pantri_test.go
@@ -1,9 +1,10 @@
-package config
+package pantri
 
 import (
 	"errors"
 	"testing"
 
+	"github.com/discentem/pantri_but_go/internal/stores"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 )
@@ -29,7 +30,7 @@ func TestDirExpanderFails(t *testing.T) {
 	conf := Config{
 		Validate: func() error { return nil },
 	}
-	dirExpander = func(path string) (string, error) {
+	expander = func(path string) (string, error) {
 		return "", errors.New("borked")
 	}
 	err := conf.Write(afero.NewMemMapFs(), "")
@@ -38,12 +39,14 @@ func TestDirExpanderFails(t *testing.T) {
 
 func TestSuccessfulWrite(t *testing.T) {
 	conf := Config{
-		Type:          "blah",
-		PantriAddress: "blahaddress",
-		Validate:      func() error { return nil },
+		Type: "blah",
+		Opts: stores.Options{
+			PantriAddress: "blahaddress",
+		},
+		Validate: func() error { return nil },
 	}
 	// override back to a dirExpander that will succeed, as opposed to previous test
-	dirExpander = func(path string) (string, error) {
+	expander = func(path string) (string, error) {
 		return path, nil
 	}
 	fsys := afero.NewMemMapFs()
