+++ internal/stores/options.go
@@ -1,0 +1,8 @@
+package stores
+
+type Options struct {
+	PantriAddress         string `mapstructure:"pantri_address"`
+	MetaDataFileExtension string `json:"metadata_file_extension" mapstructure:"metadata_file_extension"`
+	// TODO(discentem) remove this option. See #15
+	RemoveFromSourceRepo *bool `json:"remove_from_sourcerepo" mapstructure:"remove_from_sourcerepo"`
+}
