+++ internal/stores/s3/BUILD.bazel
@@ -3,18 +3,16 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "s3",
     srcs = ["s3.go"],
-    importpath = "github.com/discentem/pantri_but_go/stores/s3",
-    visibility = ["//visibility:public"],
+    importpath = "github.com/discentem/pantri_but_go/internal/stores/s3",
+    visibility = ["//:__subpackages__"],
     deps = [
-        "//metadata",
-        "//pantri",
-        "//stores",
+        "//internal/metadata",
+        "//internal/stores",
         "@com_github_aws_aws_sdk_go_v2//aws",
         "@com_github_aws_aws_sdk_go_v2_config//:config",
         "@com_github_aws_aws_sdk_go_v2_feature_s3_manager//:manager",
         "@com_github_aws_aws_sdk_go_v2_service_s3//:s3",
         "@com_github_google_logger//:logger",
-        "@com_github_mitchellh_mapstructure//:mapstructure",
         "@com_github_spf13_afero//:afero",
     ],
 )
