--- pantri/loader/BUILD.bazel
@@ -1,27 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "loader",
-    srcs = ["loader.go"],
-    importpath = "github.com/discentem/pantri_but_go/pantri/loader",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//stores",
-        "//stores/local",
-        "//stores/s3",
-        "@com_github_mitchellh_go_homedir//:go-homedir",
-        "@com_github_spf13_afero//:afero",
-    ],
-)
-
-go_test(
-    name = "loader_test",
-    size = "small",
-    timeout = "short",
-    srcs = ["loader_test.go"],
-    embed = [":loader"],
-    deps = [
-        "@com_github_spf13_afero//:afero",
-        "@com_github_stretchr_testify//assert",
-    ],
-)
