--- stores/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "stores",
-    srcs = ["stores.go"],
-    importpath = "github.com/discentem/pantri_but_go/stores",
-    visibility = ["//visibility:public"],
-    deps = ["@com_github_spf13_afero//:afero"],
-)
