--- stores/local/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "local",
-    srcs = ["local.go"],
-    importpath = "github.com/discentem/pantri_but_go/stores/local",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//metadata",
-        "//pantri",
-        "//stores",
-        "@com_github_google_logger//:logger",
-        "@com_github_mitchellh_go_homedir//:go-homedir",
-        "@com_github_mitchellh_mapstructure//:mapstructure",
-        "@com_github_spf13_afero//:afero",
-    ],
-)
